[JSORC-UPDATE]: Migrate to Process
