+++ jaseci_core/jaseci/svc/__init__.py
@@ -3,9 +3,7 @@ from .common import CommonService, ProxyService, JsOrc, MetaProperties
 from .redis import RedisService
 from .task import TaskService
 from .mail import MailService
-from .kubernetes import KubernetesService
 from .prometheus import PromotheusService
-from .jsorc import JsOrcService
 from .meta import MetaService
 
 __all__ = [
