+++ jaseci_core/jaseci/svc/config.py
@@ -2,7 +2,7 @@ KUBERNETES_CONFIG = {"in_cluster": True, "config": None}
 
 META_CONFIG = {
     "automation": False,
-    "interval": 10,
+    "backoff_interval": 10,
     "namespace": "default",
     "keep_alive": ["promon", "redis", "task", "mail"],
     "kubernetes": KUBERNETES_CONFIG,
