--- jaseci_core/jaseci/svc/jsorc/__init__.py
@@ -1,3 +1,0 @@
-from .jsorc import JsOrcService
-
-__all__ = ["JsOrcService"]
