--- jaseci_core/jaseci/svc/jsorc/config.py
@@ -1,7 +1,0 @@
-JSORC_CONFIG = {
-    "enabled": False,
-    "quiet": True,
-    "interval": 10,
-    "namespace": "default",
-    "keep_alive": ["promon", "redis", "task", "mail"],
-}
