--- jaseci_core/jaseci/svc/kubernetes/__init__.py
@@ -1,3 +1,0 @@
-from .kubernetes import KubernetesService, Kube
-
-__all__ = ["KubernetesService", "Kube"]
