--- jaseci_core/jaseci/svc/kubernetes/config.py
@@ -1,1 +1,0 @@
-KUBE_CONFIG = {"enabled": False, "quiet": True, "in_cluster": False, "config": None}
